+++ src/api/DefaultAzureResourceProvider.ts
@@ -8,9 +8,9 @@ import { getResourceGroupFromId, uiUtils } from "@microsoft/vscode-azext-azureut
 import { callWithTelemetryAndErrorHandling, getAzExtResourceType, IActionContext, nonNullProp } from '@microsoft/vscode-azext-utils';
 import { AzureResource, AzureSubscription } from '@microsoft/vscode-azext-utils/hostapi.v2';
 import * as vscode from 'vscode';
-import { AzureResourceProvider } from '../../../hostapi.v2.internal';
-import { createResourceClient } from '../../utils/azureClients';
-import { createSubscriptionContext } from '../../utils/v2/credentialsUtils';
+import { AzureResourceProvider } from '../../hostapi.v2.internal';
+import { createResourceClient } from '../utils/azureClients';
+import { createSubscriptionContext } from '../utils/v2/credentialsUtils';
 
 export class DefaultAzureResourceProvider implements AzureResourceProvider {
     private readonly onDidChangeResourceEmitter = new vscode.EventEmitter<AzureResource | undefined>();
