+++ src/api/ResourceProviderManagers.ts
@@ -5,7 +5,7 @@
 
 import { AzureResource, AzureSubscription, ResourceBase, ResourceProvider, WorkspaceResource, WorkspaceResourceProvider } from '@microsoft/vscode-azext-utils/hostapi.v2';
 import * as vscode from 'vscode';
-import { AzureResourceProvider } from '../../../hostapi.v2.internal';
+import { AzureResourceProvider } from '../../hostapi.v2.internal';
 
 export function isArray<T>(maybeArray: T[] | null | undefined): maybeArray is T[] {
     return Array.isArray(maybeArray);
