+++ src/api/compatibility/AzureResourceGroupsExtensionApi.ts
