+++ src/api/compatibility/application/CompatibleApplicationResourceTreeItem.ts
@@ -7,7 +7,7 @@ import { AzExtParentTreeItem, AzExtResourceType, AzExtTreeDataProvider, AzExtTre
 import type { ResolvedAppResourceBase } from "@microsoft/vscode-azext-utils/hostapi";
 import { AzureResource } from "@microsoft/vscode-azext-utils/hostapi.v2";
 import { TreeItemCollapsibleState } from "vscode";
-import { getIconPath } from "../../../../utils/azureUtils";
+import { getIconPath } from "../../../utils/azureUtils";
 
 /**
  * Must imitate the behavior of ResolvedAppResourceBase
