+++ src/commands/deleteResourceGroup/v2/deleteResourceGroupV2.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AzureWizard, IActionContext, UserCancelledError } from '@microsoft/vscode-azext-utils';
-import { GroupingItem } from '../../../tree/v2/azure/GroupingItem';
+import { GroupingItem } from '../../../tree/azure/GroupingItem';
 import { createActivityContext } from '../../../utils/activityUtils';
 import { localize } from '../../../utils/localize';
 import { settingUtils } from '../../../utils/settingUtils';
