+++ src/extension.ts
@@ -13,28 +13,28 @@ import * as vscode from 'vscode';
 import { AzureResourcesApiInternal } from '../hostapi.v2.internal';
 import { ActivityLogTreeItem } from './activityLog/ActivityLogsTreeItem';
 import { registerActivity } from './activityLog/registerActivity';
-import { InternalAzureResourceGroupsExtensionApi } from './api/AzureResourceGroupsExtensionApi';
-import { pickAppResource } from './api/pickAppResource';
-import { registerApplicationResourceResolver } from './api/registerApplicationResourceResolver';
-import { registerWorkspaceResourceProvider } from './api/registerWorkspaceResourceProvider';
-import { CompatibleAzExtTreeDataProvider } from './api/v2/compatibility/CompatibleAzExtTreeDataProvider';
-import { createAzureResourcesHostApi } from './api/v2/createAzureResourcesHostApi';
-import { createWrappedAzureResourcesExtensionApi } from './api/v2/createWrappedAzureResourcesExtensionApi';
-import { DefaultAzureResourceProvider } from './api/v2/DefaultAzureResourceProvider';
-import { ResourceGroupsExtensionManager } from './api/v2/ResourceGroupsExtensionManager';
-import { AzureResourceProviderManager, WorkspaceResourceProviderManager } from './api/v2/ResourceProviderManagers';
+import { InternalAzureResourceGroupsExtensionApi } from './api/compatibility/AzureResourceGroupsExtensionApi';
+import { CompatibleAzExtTreeDataProvider } from './api/compatibility/CompatibleAzExtTreeDataProvider';
+import { pickAppResource } from './api/compatibility/pickAppResource';
+import { registerApplicationResourceResolver } from './api/compatibility/registerApplicationResourceResolver';
+import { registerWorkspaceResourceProvider } from './api/compatibility/registerWorkspaceResourceProvider';
+import { createAzureResourcesHostApi } from './api/createAzureResourcesHostApi';
+import { createWrappedAzureResourcesExtensionApi } from './api/createWrappedAzureResourcesExtensionApi';
+import { DefaultAzureResourceProvider } from './api/DefaultAzureResourceProvider';
+import { ResourceGroupsExtensionManager } from './api/ResourceGroupsExtensionManager';
+import { AzureResourceProviderManager, WorkspaceResourceProviderManager } from './api/ResourceProviderManagers';
 import { registerCommands } from './commands/registerCommands';
 import { registerTagDiagnostics } from './commands/tags/registerTagDiagnostics';
 import { TagFileSystem } from './commands/tags/TagFileSystem';
 import { ext } from './extensionVariables';
+import { AzureResourceBranchDataProviderManager } from './tree/azure/AzureResourceBranchDataProviderManager';
+import { DefaultAzureResourceBranchDataProvider } from './tree/azure/DefaultAzureResourceBranchDataProvider';
+import { registerAzureTree } from './tree/azure/registerAzureTree';
 import { HelpTreeItem } from './tree/HelpTreeItem';
-import { AzureResourceBranchDataProviderManager } from './tree/v2/azure/AzureResourceBranchDataProviderManager';
-import { DefaultAzureResourceBranchDataProvider } from './tree/v2/azure/DefaultAzureResourceBranchDataProvider';
-import { registerAzureTree } from './tree/v2/azure/registerAzureTree';
-import { ResourceGroupsItem } from './tree/v2/ResourceGroupsItem';
-import { registerWorkspaceTree } from './tree/v2/workspace/registerWorkspaceTree';
-import { WorkspaceDefaultBranchDataProvider } from './tree/v2/workspace/WorkspaceDefaultBranchDataProvider';
-import { WorkspaceResourceBranchDataProviderManager } from './tree/v2/workspace/WorkspaceResourceBranchDataProviderManager';
+import { ResourceGroupsItem } from './tree/ResourceGroupsItem';
+import { registerWorkspaceTree } from './tree/workspace/registerWorkspaceTree';
+import { WorkspaceDefaultBranchDataProvider } from './tree/workspace/WorkspaceDefaultBranchDataProvider';
+import { WorkspaceResourceBranchDataProviderManager } from './tree/workspace/WorkspaceResourceBranchDataProviderManager';
 
 export async function activateInternal(context: vscode.ExtensionContext, perfStats: { loadStartTime: number; loadEndTime: number }, ignoreBundle?: boolean): Promise<AzureExtensionApiProvider> {
     ext.context = context;
