+++ src/tree/ResourceBranchDataProviderManagerBase.ts
