+++ src/tree/azure/AzureResourceBranchDataProviderManager.ts
