+++ src/tree/azure/AzureResourceGroupingManager.ts
@@ -6,13 +6,13 @@
 import { AzExtResourceType, nonNullValue, TreeItemIconPath } from '@microsoft/vscode-azext-utils';
 import { AzureResource } from '@microsoft/vscode-azext-utils/hostapi.v2';
 import * as vscode from 'vscode';
-import { azureExtensions } from '../../../azureExtensions';
-import { GroupBySettings } from '../../../commands/explorer/groupBy';
-import { ext } from '../../../extensionVariables';
-import { getIconPath, getName } from '../../../utils/azureUtils';
-import { localize } from "../../../utils/localize";
-import { settingUtils } from '../../../utils/settingUtils';
-import { treeUtils } from '../../../utils/treeUtils';
+import { azureExtensions } from '../../azureExtensions';
+import { GroupBySettings } from '../../commands/explorer/groupBy';
+import { ext } from '../../extensionVariables';
+import { getIconPath, getName } from '../../utils/azureUtils';
+import { localize } from "../../utils/localize";
+import { settingUtils } from '../../utils/settingUtils';
+import { treeUtils } from '../../utils/treeUtils';
 import { ResourceGroupsItem } from '../ResourceGroupsItem';
 import { ResourceGroupsTreeContext } from '../ResourceGroupsTreeContext';
 import { GroupingItem, GroupingItemFactory } from './GroupingItem';
