+++ src/tree/azure/DefaultAzureResourceBranchDataProvider.ts
