+++ src/tree/azure/GroupingItem.ts
@@ -7,9 +7,9 @@ import { OpenInPortalOptions } from '@microsoft/vscode-azext-azureutils';
 import { AzExtResourceType, createContextValue, ISubscriptionContext, TreeItemIconPath } from '@microsoft/vscode-azext-utils';
 import { AzureResource, AzureResourceBranchDataProvider, AzureResourceModel, AzureSubscription, ViewPropertiesModel } from '@microsoft/vscode-azext-utils/hostapi.v2';
 import * as vscode from 'vscode';
-import { ITagsModel, ResourceTags } from '../../../commands/tags/TagFileSystem';
-import { ext } from '../../../extensionVariables';
-import { getIconPath } from '../../../utils/azureUtils';
+import { ITagsModel, ResourceTags } from '../../commands/tags/TagFileSystem';
+import { ext } from '../../extensionVariables';
+import { getIconPath } from '../../utils/azureUtils';
 import { BranchDataItemOptions } from '../BranchDataProviderItem';
 import { ResourceGroupsItem } from '../ResourceGroupsItem';
 import { ResourceGroupsTreeContext } from '../ResourceGroupsTreeContext';
