+++ src/tree/workspace/WorkspaceResourceBranchDataProviderManager.ts
