+++ src/tree/workspace/registerWorkspaceTree.ts
@@ -5,13 +5,13 @@
 
 import { AzExtTreeItem } from '@microsoft/vscode-azext-utils';
 import * as vscode from 'vscode';
-import { WorkspaceResourceProviderManager } from '../../../api/v2/ResourceProviderManagers';
-import { ext } from '../../../extensionVariables';
+import { WorkspaceResourceProviderManager } from '../../api/ResourceProviderManagers';
+import { ext } from '../../extensionVariables';
 import { BranchDataItemCache } from '../BranchDataItemCache';
 import { createTreeView } from '../createTreeView';
 import { ResourceGroupsItem } from '../ResourceGroupsItem';
 import { wrapTreeForVSCode } from '../wrapTreeForVSCode';
-import { localize } from './../../../utils/localize';
+import { localize } from './../../utils/localize';
 import { WorkspaceResourceBranchDataProviderManager } from './WorkspaceResourceBranchDataProviderManager';
 import { WorkspaceResourceTreeDataProvider } from './WorkspaceResourceTreeDataProvider';
