+++ src/tree/wrapTreeForVSCode.ts
